+++ Interface/Meshtastic_Interface.py
@@ -1,5 +1,4 @@
 # MIT License - Copyright (c) 2024 Mark Qvist / unsigned.io
-
 # This example illustrates creating a custom interface
 # definition, that can be loaded and used by Reticulum at
 # runtime. Any number of custom interfaces can be created
@@ -14,12 +13,11 @@
 # port = / dev / ttyUSB0
 # speed = 115200
 
-from time import sleep
+import os
 import sys
+import struct
 import threading
 import time
-
-from packet_handling import PacketHandler
 
 
 # Let's define our custom interface class. It must
@@ -151,7 +149,6 @@ class MeshtasticInterface(Interface):
         thread.daemon = True
         thread.start()
         self.online = True
-        RNS.log("Serial port " + self.port + " is now open", RNS.LOG_VERBOSE)
 
     # This method will be called from our read-loop
     # whenever a full packet has been received over
@@ -175,16 +172,14 @@ class MeshtasticInterface(Interface):
             self.packet_index += 1
             self.packet_index = self.packet_index % 256
 
-            # Update the transmitted bytes counter
-            # and ensure that all data was written
-            self.txb += len(data)
-
     def process_message(self, packet, interface):
         """Process meshtastic traffic incoming to system"""
-        # RNS.log(packet)
+        RNS.log(packet)
         pass
 
     def write_loop(self):
+        """Writes packets from queue to meshtastic device"""
+        RNS.log('outgoing loop started')
         import meshtastic
         while True:
             data = None
@@ -194,6 +189,9 @@ class MeshtasticInterface(Interface):
                 if not data:
                     self.packet_queue.pop(0)
             if data:
+                # Update the transmitted bytes counter
+                # and ensure that all data was written
+                self.txb += len(data) - 2  # -2 for overhead
                 RNS.log(f'Sending: {data}')
                 self.interface.sendData(data,
                                    portNum=meshtastic.portnums_pb2.PRIVATE_APP,
@@ -206,7 +204,7 @@ class MeshtasticInterface(Interface):
     def connection_complete(self, interface):
         """Process meshtastic connection opened"""
         RNS.log("Connected")
-        self.online = True
+        self.configure_device()
         self.interface = interface
 
     def connection_closed(self, interface):
@@ -225,6 +223,78 @@ class MeshtasticInterface(Interface):
     # is printed in logs or external programs.
     def __str__(self):
         return "MeshtasticInterface[" + self.name + "]"
+
+class PacketHandler:
+    struct_format = 'Bb'
+    def __init__(self, data=None, index=None, max_payload=200):
+        self.max_payload = max_payload
+        self.index = index
+        self.data_dict = {}
+        self.loop_pos = 0
+        self.done = False
+        if data:  # Means we are sending
+            self.split_data(data)
+
+    def split_data(self, data: bytes):
+        """Split data into even chunks and add metadata to it"""
+        data_list = []
+        data_len = len(data)
+        num_packets = data_len//self.max_payload + 1  # Number of packets needed to hold the data
+        packet_size = data_len//num_packets + 1
+        for i in range(0, data_len, packet_size):
+            data_list.append(data[i:i + packet_size])
+        for i, packet in enumerate(data_list):
+            pos = i
+            if i == len(data_list)-1:
+                pos = -i
+            meta_data = struct.pack(self.struct_format, self.index, pos)
+            self.data_dict[i] = meta_data+packet
+
+    def get_next(self):
+        """get next packet to send"""
+        ret = self.get_index(self.loop_pos)
+        self.loop_pos += 1
+        if max(self.data_dict.keys()) < self.loop_pos:
+            self.loop_pos = 0
+            self.done = True
+        return ret
+
+    def get_index(self, i):
+        """Get the packet at an index"""
+        if i in self.data_dict:
+            return self.data_dict[i]
+
+    def is_done(self):
+        return self.done
+
+    def process_packet(self, packet: bytes):
+        """Returns data if the packet is complete, and nothing if it isn't"""
+        meta_data = packet[:len(self.struct_format)]
+        new_index, pos = struct.unpack(self.struct_format, meta_data)
+        self.index = new_index
+        self.data_dict[abs(pos)] = packet
+        if pos < 0:
+            return self.assemble_data()
+        return None
+
+    def check_data(self):
+        """check content of data dict against the expected content"""
+        expected = 0
+        for key in sorted(self.data_dict.keys()):
+            if key != expected:
+                return False
+            expected += 1
+        return True
+
+    def assemble_data(self):
+        """Put all the data together and return it or nothing if it fails"""
+        if self.check_data():
+            data = b''
+            for key in sorted(self.data_dict.keys()):
+                data += self.data_dict[key][len(self.struct_format):]
+            return data
+        else:
+            return None
 
 
 # Finally, register the defined interface class as the
