+++ Interface/packet_handling_test.py
@@ -1,0 +1,10 @@
+from Meshtastic_Interface import PacketHandler
+
+
+if __name__ == '__main__':
+    handler = PacketHandler(b'hello_world'*56, 0)
+    receive = PacketHandler()
+    while not handler.is_done():
+        packet = handler.get_next()
+        result = receive.process_packet(packet)
+    print(result == b'hello_world'*56)
